Move around Bottom Links
Bottom Links appear at the bottom of the Directory, instead of hovering on the left side.
For generated info, changed to "Streams updated at" for ease-of-understanding, and shortened to just show local time instead of date+time
