+++ src/client/Directory/Directory.css
@@ -56,3 +56,8 @@
   margin-left: 10px;
   color: var(--dark);
 }
+
+.bottomLinks * {
+  padding: 4px;
+  margin: 4px;
+}
